remove new files for unified and jax, make changes in file. still need to fix import dependencies
