+++ skbio/stats/ordination/_principal_coordinate_analysis.py
@@ -11,6 +11,14 @@ from warnings import warn
 
 import numpy as np
 import pandas as pd
+
+# Dont want dependency. - We want to remove but what if xp != np
+import cupy as cp
+import jax
+import jax.numpy as jnp
+
+
+from scipy.linalg import eigh as scipy_eigh
 from numpy import dot, hstack
 from numpy.linalg import qr, svd
 from scipy.linalg import eigh
@@ -19,12 +27,61 @@ from skbio.util import get_rng
 from skbio.stats.distance import DistanceMatrix
 from skbio.table._tabular import _create_table, _create_table_1d
 from ._ordination_results import OrdinationResults
-from ._utils import center_distance_matrix, scale
+from ._utils import center_distance_matrix as center_distance_matrix_np, scale
 from skbio.binaries import (
     pcoa_fsvd_available as _skbb_pcoa_fsvd_available,
     pcoa_fsvd as _skbb_pcoa_fsvd,
 )
 from skbio.util._decorator import params_aliased
+from skbio.util._array import ingest_array  # CL MODIFICATION
+
+
+# helpers
+def e_matrix(distance_matrix):
+    return distance_matrix * distance_matrix / -2
+
+
+def f_matrix(E_matrix):
+    row_means = E_matrix.mean(axis=1, keepdims=True)
+    col_means = E_matrix.mean(axis=0, keepdims=True)
+    matrix_mean = E_matrix.mean()
+    return E_matrix - row_means - col_means + matrix_mean
+
+
+def center_distance_matrix(distance_matrix, inplace=False):
+    """Center a distance matrix.
+
+    Note: For JAX arrays (immutable) and CuPy arrays (GPU), the ``inplace``
+    argument is accepted for API compatibility but ignored — the function
+    always returns a centered array.
+    """
+    xp, distance_matrix = ingest_array(distance_matrix)  # CL MODIFICATION
+    if xp is not np:  # CL MODIFICATION
+        return f_matrix(e_matrix(distance_matrix))  # CL MODIFICATION
+    return center_distance_matrix_np(distance_matrix, inplace=inplace)
+
+
+def _host_partial_eigh(matrix_any, subidx):
+    """Compute partial eigendecomposition on host via SciPy LAPACK.
+
+    Moves `matrix_any` to host (NumPy), ensures contiguous float64 layout,
+    runs SciPy's `eigh` with `subset_by_index`, then moves results back to
+    device for JAX/CuPy.
+    """
+    xp, matrix_any = ingest_array(matrix_any)  # CL MODIFICATION
+    # Note: asnumpy() is CuPy-specific, but asarray() works for all
+    # if xp is cp:  #CL MODIFICATION
+    #     mat_np = xp.asnumpy(matrix_any)  #CL MODIFICATION
+    # else:  #CL MODIFICATION
+    #     mat_np = np.asarray(matrix_any)  #CL MODIFICATION
+    mat_np = np.asarray(matrix_any)  # CL MODIFICATION - no cp
+    if mat_np.dtype != np.float64:
+        mat_np = mat_np.astype(np.float64, copy=False)
+    mat_np = np.ascontiguousarray(mat_np)
+
+    eigvals_np, eigvecs_np = scipy_eigh(mat_np, subset_by_index=subidx)
+
+    return xp.asarray(eigvals_np), xp.asarray(eigvecs_np)  # CL MODIFICATION
 
 
 @params_aliased(
@@ -207,7 +264,13 @@ def pcoa(
                 )
             ndim = matrix_data.shape[0]
         subidx = [matrix_data.shape[0] - ndim, matrix_data.shape[0] - 1]
-        eigvals, eigvecs = eigh(matrix_data, subset_by_index=subidx)
+        xp, matrix_data = ingest_array(matrix_data)  # CL MODIFICATION
+        if ndim < matrix_data.shape[0]:  # CL MODIFICATION
+            eigvals, eigvecs = _host_partial_eigh(
+                matrix_data, subidx
+            )  # CL MODIFICATION
+        else:  # CL MODIFICATION
+            eigvals, eigvecs = xp.linalg.eigh(matrix_data)  # CL MODIFICATION
     elif method == "fsvd":
         long_method_name = "Approximate Principal Coordinate Analysis using FSVD"
         if 0 < dimensions < 1:
@@ -229,6 +292,18 @@ def pcoa(
                 eigvals, coordinates, proportion_explained = _skbb_pcoa_fsvd(
                     distmat.data, dimensions, inplace, seed
                 )
+                xp_eigvals, eigvals = ingest_array(eigvals)  # CL MODIFICATION
+                xp_coordinates, coordinates = ingest_array(
+                    coordinates
+                )  # CL MODIFICATION
+                xp_proportion_explained, proportion_explained = ingest_array(
+                    proportion_explained
+                )  # CL MODIFICATION
+                eigvals, coordinates, proportion_explained = (
+                    xp_eigvals.asarray(eigvals),
+                    xp_coordinates.asarray(coordinates),
+                    xp_proportion_explained.asarray(proportion_explained),
+                )  # CL MODIFICATION
                 return _encapsulate_pcoa_result(
                     long_method_name,
                     eigvals,
@@ -259,8 +334,14 @@ def pcoa(
     # abs value, but that doesn't seem to be an approach accepted
     # by L&L to deal with negative eigenvalues. We raise a warning
     # in that case. First, we make values close to 0 equal to 0.
-    negative_close_to_zero = np.isclose(eigvals, 0)
-    eigvals[negative_close_to_zero] = 0
+    xp, eigvals = ingest_array(eigvals)  # CL MODIFICATION
+    negative_close_to_zero = xp.isclose(eigvals, 0)  # CL MODIFICATION
+    if xp is jnp:  # CL MODIFICATION
+        eigvals = eigvals.at[negative_close_to_zero].set(0)  # CL MODIFICATION
+    elif xp is not np:  # CL MODIFICATION
+        eigvals = xp.where(negative_close_to_zero, 0, eigvals)  # CL MODIFICATION
+    else:  # CL MODIFICATION
+        eigvals[negative_close_to_zero] = 0  # CL MODIFICATION
 
     # eigvals might not be ordered, so we first sort them, then analogously
     # sort the eigenvectors by the ordering of the eigenvalues too
@@ -285,9 +366,19 @@ def pcoa(
     # won't work as it expects all the OrdinationResults to have the same
     # number of coordinates. In order to solve this issue, we return the
     # coordinates that have a negative eigenvalue as 0
-    num_positive = (eigvals >= 0).sum()
-    eigvecs[:, num_positive:] = np.zeros(eigvecs[:, num_positive:].shape)
-    eigvals[num_positive:] = np.zeros(eigvals[num_positive:].shape)
+    xp_vec, eigvecs = ingest_array(eigvecs)  # CL MODIFICATION
+    # assume same backend for eigvals and eigvecs
+    num_positive = (eigvals >= 0).sum()  # CL MODIFICATION
+    if xp is jnp:  # CL MODIFICATION
+        eigvecs = eigvecs.at[:, num_positive:].set(0)  # CL MODIFICATION
+        eigvals = eigvals.at[num_positive:].set(0)  # CL MODIFICATION
+    else:  # CL MODIFICATION
+        eigvecs[:, num_positive:] = xp.zeros(
+            eigvecs[:, num_positive:].shape
+        )  # CL MODIFICATION
+        eigvals[num_positive:] = xp.zeros(
+            eigvals[num_positive:].shape
+        )  # CL MODIFICATION
 
     if ndim != distmat.data.shape[0]:
         # Since the dimension parameter, hereafter referred to as 'd',
@@ -302,15 +393,21 @@ def pcoa(
         # An alternative method of calculating th sum of eigenvalues is by
         # computing the trace of the centered distance matrix.
         # See proof outlined here: https://goo.gl/VAYiXx
-        sum_eigenvalues = np.trace(matrix_data)
+        xp_mat, matrix_data = ingest_array(matrix_data)  # CL MODIFICATION
+        sum_eigenvalues = xp_mat.trace(matrix_data)  # CL MODIFICATION
     else:
         # Calculate proportions the usual way
-        sum_eigenvalues = np.sum(eigvals)
+        sum_eigenvalues = xp.sum(eigvals)  # CL MODIFICATION
 
     proportion_explained = eigvals / sum_eigenvalues
     if 0 < dimensions < 1:
-        cumulative_variance = np.cumsum(proportion_explained)
-        ndim = np.searchsorted(cumulative_variance, dimensions, side="left") + 1
+        xp_prop, proportion_explained = ingest_array(
+            proportion_explained
+        )  # CL MODIFICATION
+        cumulative_variance = xp_prop.cumsum(proportion_explained)  # CL MODIFICATION
+        ndim = (
+            xp_prop.searchsorted(cumulative_variance, dimensions, side="left") + 1
+        )  # CL MODIFICATION
         # gives the number of dimensions needed to reach specified variance
         # updates number of dimensions to reach the requirement of variance.
         dimensions = ndim
@@ -329,7 +426,7 @@ def pcoa(
     # objects in the space of principal coordinates. Note that at
     # least one eigenvalue is zero because only n-1 axes are
     # needed to represent n points in a euclidean space.
-    coordinates = eigvecs * np.sqrt(eigvals)
+    coordinates = eigvecs * xp.sqrt(eigvals)  # CL MODIFICATION
 
     return _encapsulate_pcoa_result(
         long_method_name,
@@ -470,47 +567,58 @@ def _fsvd(centered_distance_matrix, dimensions=10, seed=None):
     rng = get_rng(seed)
     G = rng.standard_normal(size=(n, k))
 
+    xp, centered_distance_matrix = ingest_array(
+        centered_distance_matrix
+    )  # CL MODIFICATION
     # `use_power_method` is constantly False, so `if` won't start.
     if use_power_method:  # pragma: no cover
         # use only the given exponent
-        H = dot(centered_distance_matrix, G)
+        H = xp.dot(centered_distance_matrix, G)
 
         for x in range(2, num_levels + 2):
             # enhance decay of singular values
             # note: distance_matrix is no longer transposed, saves work
             # since we're expecting symmetric, square matrices anyway
             # (Daniel McDonald's changes)
-            H = dot(centered_distance_matrix, dot(centered_distance_matrix, H))
+            H = xp.dot(centered_distance_matrix, xp.dot(centered_distance_matrix, H))
 
     else:
         # compute the m x l matrices H^{(0)}, ..., H^{(i)}
         # Note that this is done implicitly in each iteration below.
-        H = dot(centered_distance_matrix, G)
+        H = xp.dot(centered_distance_matrix, G)
         # to enhance performance
-        H = hstack((H, dot(centered_distance_matrix, dot(centered_distance_matrix, H))))
+        H = hstack(
+            (H, xp.dot(centered_distance_matrix, xp.dot(centered_distance_matrix, H)))
+        )
 
         # `num_levels` is constantly 1, so `for` loop won't start
         for x in range(3, num_levels + 2):  # pragma: no cover
-            tmp = dot(centered_distance_matrix, dot(centered_distance_matrix, H))
+            tmp = xp.dot(centered_distance_matrix, xp.dot(centered_distance_matrix, H))
 
             H = hstack(
-                (H, dot(centered_distance_matrix, dot(centered_distance_matrix, tmp)))
+                (
+                    H,
+                    xp.dot(
+                        centered_distance_matrix, xp.dot(centered_distance_matrix, tmp)
+                    ),
+                )
             )
 
     # Using the pivoted QR-decomposition, form a real m * ((i+1)l) matrix Q
     # whose columns are orthonormal, s.t. there exists a real
     # ((i+1)l) * ((i+1)l) matrix R for which H = QR
-    Q, R = qr(H)
+    xp_q, H = ingest_array(H)  # CL MODIFICATION
+    Q, R = xp_q.qr(H)
 
     # Compute the n * ((i+1)l) product matrix T = A^T Q
-    T = dot(centered_distance_matrix, Q)  # step 3
-
+    T = xp.dot(centered_distance_matrix, Q)  # step 3
     # Form an SVD of T
-    Vt, St, W = svd(T, full_matrices=False)
+    xp_s, T = ingest_array(T)  # CL MODIFICATION
+    Vt, St, W = xp_s.svd(T, full_matrices=False)
     W = W.transpose()
 
     # Compute the m * ((i+1)l) product matrix
-    Ut = dot(Q, W)
+    Ut = xp.dot(Q, W)
 
     U_fsvd = Ut[:, :dimensions]
 
@@ -579,15 +687,18 @@ def pcoa_biplot(ordination, y):
     # S_pc from equation 9.44
     # Represents the covariance matrix between the features matrix and the
     # column-centered eigenvectors of the pcoa.
-    spc = (1 / (N - 1)) * y.values.T.dot(scale(coordinates, ddof=1))
+    xp, coordinates = ingest_array(coordinates)  # CL MODIFICATION
+    spc = (1 / (N - 1)) * xp.dot(
+        y.values.T, scale(coordinates, ddof=1)
+    )  # CL MODIFICATION
 
     # U_proj from equation 9.55, is the matrix of descriptors to be projected.
     #
     # Only get the power of non-zero values, otherwise this will raise a
     # divide by zero warning. There shouldn't be negative eigenvalues(?)
-    Uproj = np.sqrt(N - 1) * spc.dot(
-        np.diag(np.power(eigvals, -0.5, where=eigvals > 0))
-    )
+    Uproj = xp.sqrt(N - 1) * xp.dot(
+        spc, xp.diag(xp.power(eigvals, -0.5, where=eigvals > 0))
+    )  # CL MODIFICATION
 
     ordination.features = pd.DataFrame(
         data=Uproj, index=y.columns.copy(), columns=coordinates.columns.copy()
